Variable type fix

+++ src/cryptonote_core/cryptonote_tx_utils.cpp
@@ -100,7 +100,7 @@ keypair get_deterministic_keypair_from_height(uint64_t height)
   return k;
 }
 
-uint64_t get_governance_reward(uint64_t height, uint64_t base_reward, uint8_t hf_version, uint8_t hardfork_height)
+uint64_t get_governance_reward(uint64_t height, uint64_t base_reward, uint8_t hf_version, uint64_t hardfork_height)
 {
   if(hf_version >= 17 && height == hardfork_height)
 	  return DEVS_REWARD_V17;
